moved helpers from conftest to common
