+++ tests/conftest.py
@@ -11,16 +11,8 @@ from homeassistant.components.binary_sensor import (
     DOMAIN as BINARY_SENSOR_DOMAIN,
     BinarySensorDeviceClass,
 )
-from homeassistant.config_entries import ConfigEntryState
-from homeassistant.const import ATTR_FLOOR_ID, CONF_PLATFORM
 from homeassistant.core import HomeAssistant
-from homeassistant.helpers.area_registry import async_get as async_get_ar
-from homeassistant.helpers.entity_registry import async_get as async_get_er
-from homeassistant.helpers.floor_registry import async_get as async_get_fr
-from homeassistant.setup import async_setup_component
 
-from .const import DEFAULT_MOCK_AREA, MOCK_AREAS, MockAreaIds
-from .mocks import MockBinarySensor
 from custom_components.magic_areas.const import (
     CONF_AGGREGATES_MIN_ENTITIES,
     CONF_ENABLED_FEATURES,
@@ -30,98 +22,16 @@ from custom_components.magic_areas.const import (
     AreaType,
 )
 
-from tests.common import get_basic_config_entry_data, setup_test_component_platform
+from tests.common import (
+    get_basic_config_entry_data,
+    init_integration,
+    setup_mock_entities,
+    shutdown_integration,
+)
+from tests.const import DEFAULT_MOCK_AREA, MOCK_AREAS, MockAreaIds
+from tests.mocks import MockBinarySensor
 
 _LOGGER = logging.getLogger(__name__)
-
-# Helpers
-
-
-async def init_integration(
-    hass: HomeAssistant,
-    config_entries: list[MockConfigEntry],
-    areas: list[MockAreaIds] | None = None,
-) -> None:
-    """Set up the integration."""
-
-    if not areas:
-        areas = [DEFAULT_MOCK_AREA]
-
-    area_registry = async_get_ar(hass)
-    floor_registry = async_get_fr(hass)
-
-    # Register areas
-    for area in areas:
-        area_object = MOCK_AREAS[area]
-        floor_id: str | None = None
-
-        if area_object[ATTR_FLOOR_ID]:
-            assert area_object[ATTR_FLOOR_ID] is not None
-            floor_name = str(area_object[ATTR_FLOOR_ID])
-            floor_entry = floor_registry.async_get_floor_by_name(floor_name)
-            if not floor_entry:
-                floor_entry = floor_registry.async_create(floor_name)
-            assert floor_entry is not None
-            floor_id = floor_entry.floor_id
-        area_registry.async_create(name=area.value, floor_id=floor_id)
-
-    for config_entry in config_entries:
-        config_entry.add_to_hass(hass)
-
-    assert await async_setup_component(hass, DOMAIN, {})
-    await hass.async_block_till_done()
-
-    for config_entry in config_entries:
-        assert config_entry.state is ConfigEntryState.LOADED
-
-
-async def shutdown_integration(
-    hass: HomeAssistant, config_entries: list[MockConfigEntry]
-) -> None:
-    """Teardown the integration."""
-
-    _LOGGER.info("Unloading integration.")
-    for config_entry in config_entries:
-        await hass.config_entries.async_unload(config_entry.entry_id)
-        await hass.async_block_till_done()
-
-    assert not hass.data.get(DOMAIN)
-
-    for config_entry in config_entries:
-        assert config_entry.state is ConfigEntryState.NOT_LOADED
-    _LOGGER.info("Integration unloaded.")
-
-
-async def setup_mock_entities(
-    hass: HomeAssistant, domain: str, area_entity_map: dict[MockAreaIds, list[Any]]
-) -> None:
-    """Set up multiple mock entities at once."""
-
-    all_entities: list[Any] = []
-    entity_area_map: dict[Any, MockAreaIds] = {}
-
-    for area_id, entity_list in area_entity_map.items():
-        for entity in entity_list:
-            all_entities.append(entity)
-            entity_area_map[entity.unique_id] = area_id
-
-    # Setup entities
-    setup_test_component_platform(hass, domain, all_entities)
-    assert await async_setup_component(hass, domain, {domain: {CONF_PLATFORM: "test"}})
-    await hass.async_block_till_done()
-
-    # Update area IDs
-    entity_registry = async_get_er(hass)
-    for entity in all_entities:
-        assert entity is not None
-        assert entity.entity_id is not None
-        assert entity.unique_id is not None
-        entity_registry.async_update_entity(
-            entity.entity_id,
-            area_id=entity_area_map[entity.unique_id].value,
-        )
-    await hass.async_block_till_done()
-
 
 # Fixtures
