+++ tests/test_threshold.py
@@ -17,14 +17,6 @@ from homeassistant.components.threshold.const import ATTR_HYSTERESIS, ATTR_UPPER
 from homeassistant.const import LIGHT_LUX, STATE_OFF, STATE_ON
 from homeassistant.core import HomeAssistant
 
-from .common import get_basic_config_entry_data
-from .conftest import (
-    DEFAULT_MOCK_AREA,
-    init_integration,
-    setup_mock_entities,
-    shutdown_integration,
-)
-from .mocks import MockBinarySensor, MockSensor
 from custom_components.magic_areas.const import (
     CONF_AGGREGATES_ILLUMINANCE_THRESHOLD,
     CONF_AGGREGATES_ILLUMINANCE_THRESHOLD_HYSTERESIS,
@@ -34,7 +26,17 @@ from custom_components.magic_areas.const import (
     DOMAIN,
 )
 
+from tests.common import (
+    get_basic_config_entry_data,
+    init_integration,
+    setup_mock_entities,
+    shutdown_integration,
+)
+from tests.const import DEFAULT_MOCK_AREA
+from tests.mocks import MockBinarySensor, MockSensor
+
 _LOGGER = logging.getLogger(__name__)
+
 
 # Fixtures
